+++ src/main/java/ru/hh/techradar/exception/EntityExistsException.java
@@ -1,4 +1,4 @@
-package ru.hh.techradar.entity.exception;
+package ru.hh.techradar.exception;
 
 public class EntityExistsException extends RuntimeException {
