+++ src/main/java/ru/hh/techradar/exception/NotFoundExceptionMapper.java
@@ -1,4 +1,4 @@
-package ru.hh.techradar.entity.exception;
+package ru.hh.techradar.exception;
 
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
