+++ src/main/java/ru/hh/techradar/security/dto/Constants.java
@@ -6,6 +6,5 @@ public class Constants {
 
   public final static String AUTHORIZATION = "Authorization";
   public final static String TOKEN_TYPE = "Bearer ";
-  public final static String USER_NOT_FOUND = "User not found!";
   public final static int TOKEN_TYPE_LENGTH = 7;
 }
