+++ src/main/java/ru/hh/techradar/service/BlipEventService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.hh.techradar.entity.BlipEvent;
-import ru.hh.techradar.entity.exception.NotFoundException;
+import ru.hh.techradar.exception.NotFoundException;
 import ru.hh.techradar.filter.BlipFilter;
 import ru.hh.techradar.mapper.BlipEventMapper;
 import ru.hh.techradar.repository.BlipEventRepository;
