+++ src/main/java/ru/hh/techradar/service/RadarService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.hh.techradar.entity.Radar;
-import ru.hh.techradar.entity.exception.NotFoundException;
+import ru.hh.techradar.exception.NotFoundException;
 import ru.hh.techradar.filter.ComponentFilter;
 import ru.hh.techradar.repository.RadarRepository;
