+++ src/main/java/ru/hh/techradar/service/RingSettingService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.hh.techradar.entity.RingSetting;
-import ru.hh.techradar.entity.exception.NotFoundException;
+import ru.hh.techradar.exception.NotFoundException;
 import ru.hh.techradar.repository.RingSettingRepository;
 
 @Service
