+++ src/main/java/ru/hh/techradar/service/UserService.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.hh.techradar.entity.User;
-import ru.hh.techradar.entity.exception.EntityExistsException;
-import ru.hh.techradar.entity.exception.NotFoundException;
+import ru.hh.techradar.exception.EntityExistsException;
+import ru.hh.techradar.exception.NotFoundException;
 import ru.hh.techradar.filter.UserFilter;
 import ru.hh.techradar.mapper.UserMapper;
 import ru.hh.techradar.repository.UserRepository;
